refactor(types): move domain types from api/backend to src/types/
