+++ src/components/MainPage.test.tsx
@@ -46,8 +46,6 @@ import * as connectionState from "../utils/connectionState";
 import type {
   MigrationStatus,
   SaveSortMigrationStatus,
-} from "../api/backend";
-import type {
   SyncStats,
   SyncPreview,
   SyncPreviewSummary,
