+++ src/components/saves/SlotPanel.tsx
@@ -8,8 +8,7 @@ import { useState, useRef, useEffect, createElement, FC } from "react";
 import { ConfirmModal, DialogButton, showModal } from "@decky/ui";
 import { toaster } from "@decky/api";
 import { getSlotSaves, switchSlot, debugLog, getSlotDeleteInfo, deleteSlot } from "../../api/backend";
-import type { SlotDeleteInfo } from "../../api/backend";
-import type { SaveStatus, SyncConflict, SaveSlotSummary, SlotSaveFile, SwitchSlotResponse } from "../../types";
+import type { SaveStatus, SyncConflict, SaveSlotSummary, SlotSaveFile, SwitchSlotResponse, SlotDeleteInfo } from "../../types";
 import { scrollFocusedToCenter } from "../../utils/scrollHelpers";
 import { MUTED_COLOR, computeSyncSummary, displaySlot, slotDeleteFailureToast } from "./helpers";
 import { renderSaveFileRow } from "./SaveFileRow";
